Move petstore into package pet
