+++ codegen/main.go
@@ -1,13 +1,15 @@
 package main
 
 import (
-	"github.com/labstack/echo/v4"
+	"kbtg-bootcamp-petstore/pet"
 	"kbtg-bootcamp-petstore/petstore"
+
+	"github.com/labstack/echo/v4"
 )
 
 func main() {
 	e := echo.New()
-	api := NewServerWrapper()
+	api := pet.NewServerWrapper()
 	petstore.RegisterHandlers(e, api)
 	e.Logger.Fatal(e.Start(":1323"))
 }
